+++ src/waveform/waveformwidgetfactory.cpp
@@ -399,7 +399,7 @@ bool WaveformWidgetFactory::setWaveformWidget(WWaveformViewer* viewer,
     if (index == -1) {
         // add holder
         m_waveformWidgetHolders.push_back(std::move(holder));
-        index = m_waveformWidgetHolders.size() - 1;
+        index = static_cast<int>(m_waveformWidgetHolders.size()) - 1;
     } else {
         // update holder
         DEBUG_ASSERT(index >= 0);
@@ -628,7 +628,8 @@ void WaveformWidgetFactory::notifyZoomChange(WWaveformViewer* viewer) {
 }
 
 void WaveformWidgetFactory::render() {
-    ScopedTimer t("WaveformWidgetFactory::render() %1waveforms", m_waveformWidgetHolders.size());
+    ScopedTimer t("WaveformWidgetFactory::render() %1waveforms",
+            static_cast<int>(m_waveformWidgetHolders.size()));
 
     //int paintersSetupTime0 = 0;
     //int paintersSetupTime1 = 0;
@@ -638,7 +639,8 @@ void WaveformWidgetFactory::render() {
     if (!m_skipRender) {
         if (m_type) {   // no regular updates for an empty waveform
             // next rendered frame is displayed after next buffer swap and than after VSync
-            QVarLengthArray<bool, 10> shouldRenderWaveforms(m_waveformWidgetHolders.size());
+            QVarLengthArray<bool, 10> shouldRenderWaveforms(
+                    static_cast<int>(m_waveformWidgetHolders.size()));
             for (std::size_t i = 0; i < m_waveformWidgetHolders.size(); i++) {
                 WaveformWidgetAbstract* pWaveformWidget = m_waveformWidgetHolders[i].m_waveformWidget;
                 // Don't bother doing the pre-render work if we aren't going to
@@ -694,7 +696,8 @@ void WaveformWidgetFactory::render() {
 }
 
 void WaveformWidgetFactory::swap() {
-    ScopedTimer t("WaveformWidgetFactory::swap() %1waveforms", m_waveformWidgetHolders.size());
+    ScopedTimer t("WaveformWidgetFactory::swap() %1waveforms",
+            static_cast<int>(m_waveformWidgetHolders.size()));
 
     // Do this in an extra slot to be sure to hit the desired interval
     if (!m_skipRender) {
